Refactor proposal
