+++ src/ganzo/sources/gcs.py
@@ -1,10 +1,11 @@
 import os
-from distutils.dir_util import copy_tree
 
 from google.cloud import storage
 
+from ganzo.sources import TemplateSource
 
-class GCSSource:
+
+class GCSSource(TemplateSource):
     def __init__(self, bucket_name: str):
         self.client = storage.Client()
         self.bucket_name = bucket_name
@@ -31,10 +32,3 @@ class GCSSource:
 
         if empty:
             raise ValueError(f"Template '{template_name}' does not exist.")
-
-class LocalSource:
-
-    def load_template(self, template_name: str, target_path: str):
-        template_path = f"{template_name}/"
-        print(f"Copying '{template_path}' into '{target_path}'")
-        copy_tree(template_path,target_path)
